添加重试场景
