+++ src/utils/index.ts
@@ -3,5 +3,6 @@ export * from './observable'
 
 export * from './base64'
 export * from './helper'
+export * from './config'
 
 export { default as compressImage, CompressResult } from './compress'
